Core - API - Reorganizing of utilities
