+++ software/core/oqm-core-api/src/main/java/tech/ebp/oqm/core/api/model/validation/validators/identifiers/GTIN14Validator.java
@@ -1,11 +1,9 @@
 package tech.ebp.oqm.core.api.model.validation.validators.identifiers;
 
 import jakarta.validation.ConstraintValidatorContext;
-import tech.ebp.oqm.core.api.model.validation.annotations.identifiers.ValidEAN8;
 import tech.ebp.oqm.core.api.model.validation.annotations.identifiers.ValidGTIN14;
 import tech.ebp.oqm.core.api.model.validation.validators.Validator;
-import tech.ebp.oqm.core.api.service.codes.upc.EANCodeUtilities;
-import tech.ebp.oqm.core.api.service.codes.upc.GTINCodeUtilities;
+import tech.ebp.oqm.core.api.service.identifiers.general.upc.GTINCodeUtilities;
 
 import java.util.ArrayList;
 import java.util.List;
