+++ software/core/oqm-core-api/src/main/java/tech/ebp/oqm/core/api/model/validation/validators/identifiers/ISBN13Validator.java
@@ -2,10 +2,8 @@ package tech.ebp.oqm.core.api.model.validation.validators.identifiers;
 
 import jakarta.validation.ConstraintValidatorContext;
 import tech.ebp.oqm.core.api.model.validation.annotations.identifiers.ValidISBN13;
-import tech.ebp.oqm.core.api.model.validation.annotations.identifiers.ValidUPCE;
 import tech.ebp.oqm.core.api.model.validation.validators.Validator;
-import tech.ebp.oqm.core.api.service.codes.upc.ISBNCodeUtilities;
-import tech.ebp.oqm.core.api.service.codes.upc.UpcCodeUtilities;
+import tech.ebp.oqm.core.api.service.identifiers.general.upc.ISBNCodeUtilities;
 
 import java.util.ArrayList;
 import java.util.List;
