+++ software/core/oqm-core-api/src/main/java/tech/ebp/oqm/core/api/model/validation/validators/identifiers/UPCAValidator.java
@@ -3,7 +3,7 @@ package tech.ebp.oqm.core.api.model.validation.validators.identifiers;
 import jakarta.validation.ConstraintValidatorContext;
 import tech.ebp.oqm.core.api.model.validation.annotations.identifiers.ValidUPCA;
 import tech.ebp.oqm.core.api.model.validation.validators.Validator;
-import tech.ebp.oqm.core.api.service.codes.upc.UpcCodeUtilities;
+import tech.ebp.oqm.core.api.service.identifiers.general.upc.UpcCodeUtilities;
 
 import java.util.ArrayList;
 import java.util.List;
