+++ software/core/oqm-core-api/src/main/java/tech/ebp/oqm/core/api/service/identifiers/general/upc/GTINCodeUtilities.java
@@ -1,8 +1,7 @@
 
-package tech.ebp.oqm.core.api.service.codes.upc;
+package tech.ebp.oqm.core.api.service.identifiers.general.upc;
 
 import lombok.NonNull;
-import org.apache.commons.validator.routines.checkdigit.EAN13CheckDigit;
 
 /**
  * Class to handle GTIN codes not covered by more specific formats.
